[go-gen][metrics] Extract shared logic to common generator
